Split out manifest processing
